- move the profile-picker to a floating button to save space
