+++ front/components/settings/profile-picker/profile-picker-float.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <van-floating-bubble v-model:offset="offset" axis="y" magnetic="x" @click="isListVisible = true" :gap="0" class="profile-floating-button">
+      <icon-settings size="20" color="#fff" stroke="1.6" />
+    </van-floating-bubble>
+
+    <van-popup v-model:show="isListVisible" position="bottom" :style="{ height: '40%' }">
+      <profile-picker-list></profile-picker-list>
+    </van-popup>
+  </div>
+</template>
+
+<script setup>
+import { IconSettings } from '@tabler/icons-vue'
+
+const isListVisible = ref(false)
+
+const offset = ref({ y: window.innerHeight / 2.2 })
+</script>
