+++ front/components/settings/profile-picker/profile-picker-form.vue
