+++ front/layouts/default.vue
@@ -1,6 +1,8 @@
 <template>
   <div :class="layoutClass">
     <slot />
+
+    <profile-picker-float v-if="isProfileFloatButtonVisible" />
     <app-bottom-toolbar />
     <app-bottom-loading />
   </div>
@@ -9,6 +11,8 @@
 import { ref } from 'vue';
 
 <script setup>
+import RouteConstants from '~/constants/RouteConstants.js'
+
 const device = useDevice()
 
 const layoutClass = computed(() => {
@@ -17,5 +21,10 @@ const layoutClass = computed(() => {
     desktop: device.isDesktop,
     mobile: device.isMobile,
   }
+})
+
+const isProfileFloatButtonVisible = computed(() => {
+  let route = useRoute()
+  return RouteConstants.isSettings(route.path) || route.path === RouteConstants.ROUTE_TRANSACTION_ID
 })
 </script>
