--- front/pages/mihai.vue
@@ -1,11 +1,0 @@
-<template>
-  <div>Page Mihai</div>
-</template>
-
-<script setup>
-definePageMeta({
-  layout: 'empty',
-})
-
-onMounted(() => {})
-</script>
